+++ crates/swc_ecma_minifier/tests/fixture/next/33265/static/chunks/pages/index-cb36c1bf7f830e3c/output.js
@@ -643,7 +643,7 @@
                         offset: 4
                     }) || (0, byte_helpers.G3)(bytes, CONSTANTS.fmp4, {
                         offset: 4
-                    })) || !!((0, byte_helpers.G3)(bytes, CONSTANTS.moof, {
+                    }) || (0, byte_helpers.G3)(bytes, CONSTANTS.moof, {
                         offset: 4
                     }) || (0, byte_helpers.G3)(bytes, CONSTANTS.moov, {
                         offset: 4
@@ -1198,8 +1198,7 @@
             }
             function needNamespaceDefine(node, isHTML, visibleNamespaces) {
                 var prefix = node.prefix || "", uri = node.namespaceURI;
-                if (!uri) return !1;
-                if ("xml" === prefix && uri === NAMESPACE.XML || uri === NAMESPACE.XMLNS) return !1;
+                if (!uri || "xml" === prefix && uri === NAMESPACE.XML || uri === NAMESPACE.XMLNS) return !1;
                 for(var i = visibleNamespaces.length; i--;){
                     var ns = visibleNamespaces[i];
                     if (ns.prefix === prefix) return ns.namespace !== uri;
@@ -4065,9 +4064,7 @@
                             continue;
                         }
                         var m2 = t.match(/^<([^.\s/0-9>]+)(\.[^\s\\>]+)?([^>\\]+)?(\\?)>?$/);
-                        if (!m2) continue;
-                        if (!(node = createElement(m2[1], m2[3]))) continue;
-                        if (!shouldAdd(current1, node)) continue;
+                        if (!m2 || !(node = createElement(m2[1], m2[3])) || !shouldAdd(current1, node)) continue;
                         if (m2[2]) {
                             var classes = m2[2].split(".");
                             classes.forEach(function(cl) {
@@ -5190,8 +5187,7 @@
                 if ("string" == typeof value) return fromString(value, encodingOrOffset);
                 if (ArrayBuffer.isView(value)) return fromArrayLike(value);
                 if (null == value) throw new TypeError("The first argument must be one of type string, Buffer, ArrayBuffer, Array, or Array-like Object. Received type " + typeof value);
-                if (isInstance(value, ArrayBuffer) || value && isInstance(value.buffer, ArrayBuffer)) return fromArrayBuffer(value, encodingOrOffset, length);
-                if ("undefined" != typeof SharedArrayBuffer && (isInstance(value, SharedArrayBuffer) || value && isInstance(value.buffer, SharedArrayBuffer))) return fromArrayBuffer(value, encodingOrOffset, length);
+                if (isInstance(value, ArrayBuffer) || value && isInstance(value.buffer, ArrayBuffer) || "undefined" != typeof SharedArrayBuffer && (isInstance(value, SharedArrayBuffer) || value && isInstance(value.buffer, SharedArrayBuffer))) return fromArrayBuffer(value, encodingOrOffset, length);
                 if ("number" == typeof value) throw new TypeError('The "value" argument must not be of type number. Received type number');
                 var valueOf = value.valueOf && value.valueOf();
                 if (null != valueOf && valueOf !== value) return Buffer.from(valueOf, encodingOrOffset, length);
@@ -5263,9 +5259,7 @@
             }
             function slowToString(encoding, start, end) {
                 var loweredCase = !1;
-                if ((void 0 === start || start < 0) && (start = 0), start > this.length) return "";
-                if ((void 0 === end || end > this.length) && (end = this.length), end <= 0) return "";
-                if ((end >>>= 0) <= (start >>>= 0)) return "";
+                if ((void 0 === start || start < 0) && (start = 0), start > this.length || ((void 0 === end || end > this.length) && (end = this.length), end <= 0 || (end >>>= 0) <= (start >>>= 0))) return "";
                 for(encoding || (encoding = "utf8");;)switch(encoding){
                     case "hex":
                         return hexSlice(this, start, end);
@@ -5422,8 +5416,7 @@
                 if (offset + ext > buf.length) throw new RangeError("Index out of range");
             }
             function checkIEEE754(buf, value, offset, ext, max, min) {
-                if (offset + ext > buf.length) throw new RangeError("Index out of range");
-                if (offset < 0) throw new RangeError("Index out of range");
+                if (offset + ext > buf.length || offset < 0) throw new RangeError("Index out of range");
             }
             function writeFloat(buf, value, offset, littleEndian, noAssert) {
                 return value = +value, offset >>>= 0, noAssert || checkIEEE754(buf, value, offset, 4, 3.4028234663852886e38, -340282346638528860000000000000000000000), ieee754.write(buf, value, offset, littleEndian, 23, 4), offset + 4;
@@ -5693,8 +5686,7 @@
                 return writeDouble(this, value, offset, !1, noAssert);
             }, Buffer.prototype.copy = function(target, targetStart, start, end) {
                 if (!Buffer.isBuffer(target)) throw new TypeError("argument should be a Buffer");
-                if (start || (start = 0), end || 0 === end || (end = this.length), targetStart >= target.length && (targetStart = target.length), targetStart || (targetStart = 0), end > 0 && end < start && (end = start), end === start) return 0;
-                if (0 === target.length || 0 === this.length) return 0;
+                if (start || (start = 0), end || 0 === end || (end = this.length), targetStart >= target.length && (targetStart = target.length), targetStart || (targetStart = 0), end > 0 && end < start && (end = start), end === start || 0 === target.length || 0 === this.length) return 0;
                 if (targetStart < 0) throw new RangeError("targetStart out of bounds");
                 if (start < 0 || start >= this.length) throw new RangeError("Index out of range");
                 if (end < 0) throw new RangeError("sourceEnd out of bounds");
@@ -5729,11 +5721,7 @@
                 for(var codePoint, length = string.length, leadSurrogate = null, bytes = [], i = 0; i < length; ++i){
                     if ((codePoint = string.charCodeAt(i)) > 0xd7ff && codePoint < 0xe000) {
                         if (!leadSurrogate) {
-                            if (codePoint > 0xdbff) {
-                                (units -= 3) > -1 && bytes.push(0xef, 0xbf, 0xbd);
-                                continue;
-                            }
-                            if (i + 1 === length) {
+                            if (codePoint > 0xdbff || i + 1 === length) {
                                 (units -= 3) > -1 && bytes.push(0xef, 0xbf, 0xbd);
                                 continue;
                             }
@@ -5830,8 +5818,7 @@
             var setPrototypeOf = __webpack_require__(9611);
             function _construct(Parent1, args1, Class1) {
                 return (_construct = !function() {
-                    if ("undefined" == typeof Reflect || !Reflect.construct) return !1;
-                    if (Reflect.construct.sham) return !1;
+                    if ("undefined" == typeof Reflect || !Reflect.construct || Reflect.construct.sham) return !1;
                     if ("function" == typeof Proxy) return !0;
                     try {
                         return Boolean.prototype.valueOf.call(Reflect.construct(Boolean, [], function() {})), !0;